Fix for issue #8 utils functions removal
