+++ database/db.log
@@ -1,1 +1,2 @@
 2024/07/05 06:31:34 New board created. name=web-service, active=false
+2024/07/09 06:27:50 New board created. name=f comp cli tool, active=true
